requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
frozenlist==1.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.3
python-dotenv==0.21.0
telepot==12.7
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
